requirements.txt
# Core web
flask==3.0.3
flask-cors==4.0.0

# Computer vision + YOLO
opencv-python==4.9.0.80
ultralytics==8.2.103
pillow==10.3.0

# Flask deps (pinned to avoid conflicts)
click==8.1.7
jinja2==3.1.4
markupsafe==2.1.5
itsdangerous==2.2.0
werkzeug==3.0.3
blinker==1.9.0

# Numpy compatible with OpenCV/Ultralytics on Windows
numpy==1.26.4
